+++ src/shellyDevice.realgen3.test.ts
@@ -16,7 +16,7 @@ describe('Shellies', () => {
   let device: ShellyDevice | undefined;
 
   const firmwareGen2 = '1.4.4-g6d2a586';
-  const address = '30:f6:ef:69:2b:c5';
+  const address = 'c4:cb:76:b3:cd:1f';
 
   beforeAll(async () => {
     consoleLogSpy = jest.spyOn(console, 'log').mockImplementation((...args: any[]) => {
@@ -188,26 +188,124 @@ describe('Shellies', () => {
     expect(device.getComponentIds()).toStrictEqual(['ble', 'cloud', 'input:0', 'mqtt', 'switch:0', 'sys', 'sntp', 'wifi_ap', 'wifi_sta', 'wifi_sta1', 'ws']);
 
     expect(device.bthomeTrvs.size).toBe(0);
-    expect(device.bthomeDevices.size).toBe(0);
-    expect(device.bthomeSensors.size).toBe(0);
+
+    expect(device.bthomeDevices.size).toBe(4);
+    expect(device.bthomeDevices.has('0c:ef:f6:f1:d7:7b')).toBe(true);
+    expect(device.bthomeDevices.get('0c:ef:f6:f1:d7:7b')?.model).toBe('Shelly BLU DoorWindow');
+    expect(device.bthomeDevices.has('38:39:8f:8b:d2:29')).toBe(true);
+    expect(device.bthomeDevices.get('38:39:8f:8b:d2:29')?.model).toBe('Shelly BLU Button1');
+    expect(device.bthomeDevices.has('7c:c6:b6:65:2d:87')).toBe(true);
+    expect(device.bthomeDevices.get('7c:c6:b6:65:2d:87')?.model).toBe('Shelly BLU HT');
+    expect(device.bthomeDevices.has('0c:ae:5f:5a:0b:fa')).toBe(true);
+    expect(device.bthomeDevices.get('0c:ae:5f:5a:0b:fa')?.model).toBe('Shelly BLU Motion');
+
+    expect(device.bthomeSensors.size).toBe(14);
+
+    // BLU DoorWindow
+    expect(device.bthomeSensors.has('bthomesensor:200')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:200')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:200')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:200')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:200')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:201')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:201')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:201')?.name).toBe('Illuminance');
+    expect(device.bthomeSensors.get('bthomesensor:201')?.sensorId).toBe(5);
+    expect(device.bthomeSensors.get('bthomesensor:201')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:202')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:202')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:202')?.name).toBe('Contact');
+    expect(device.bthomeSensors.get('bthomesensor:202')?.sensorId).toBe(45);
+    expect(device.bthomeSensors.get('bthomesensor:202')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:203')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:203')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:203')?.name).toBe('Rotation');
+    expect(device.bthomeSensors.get('bthomesensor:203')?.sensorId).toBe(63);
+    expect(device.bthomeSensors.get('bthomesensor:203')?.sensorIdx).toBe(0);
+
+    // BLU Button
+    expect(device.bthomeSensors.has('bthomesensor:204')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:204')?.addr).toBe('38:39:8f:8b:d2:29');
+    expect(device.bthomeSensors.get('bthomesensor:204')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:204')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:204')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:205')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:205')?.addr).toBe('38:39:8f:8b:d2:29');
+    expect(device.bthomeSensors.get('bthomesensor:205')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:205')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:205')?.sensorIdx).toBe(0);
+
+    // BLU HT
+    expect(device.bthomeSensors.has('bthomesensor:206')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:206')?.addr).toBe('7c:c6:b6:65:2d:87');
+    expect(device.bthomeSensors.get('bthomesensor:206')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:206')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:206')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:207')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:207')?.addr).toBe('7c:c6:b6:65:2d:87');
+    expect(device.bthomeSensors.get('bthomesensor:207')?.name).toBe('Humidity');
+    expect(device.bthomeSensors.get('bthomesensor:207')?.sensorId).toBe(46);
+    expect(device.bthomeSensors.get('bthomesensor:207')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:208')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:208')?.addr).toBe('7c:c6:b6:65:2d:87');
+    expect(device.bthomeSensors.get('bthomesensor:208')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:208')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:208')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:209')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:209')?.addr).toBe('7c:c6:b6:65:2d:87');
+    expect(device.bthomeSensors.get('bthomesensor:209')?.name).toBe('Temperature');
+    expect(device.bthomeSensors.get('bthomesensor:209')?.sensorId).toBe(69);
+    expect(device.bthomeSensors.get('bthomesensor:209')?.sensorIdx).toBe(0);
+
+    // BLU Motion
+    expect(device.bthomeSensors.has('bthomesensor:210')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:210')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:210')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:210')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:210')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:211')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:211')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:211')?.name).toBe('Motion');
+    expect(device.bthomeSensors.get('bthomesensor:211')?.sensorId).toBe(33);
+    expect(device.bthomeSensors.get('bthomesensor:211')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:212')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:212')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:212')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:212')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:212')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:213')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:213')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:213')?.name).toBe('Illuminance');
+    expect(device.bthomeSensors.get('bthomesensor:213')?.sensorId).toBe(5);
+    expect(device.bthomeSensors.get('bthomesensor:213')?.sensorIdx).toBe(0);
 
     const component = device.getComponent('switch:0');
     expect(component).not.toBeUndefined();
 
     // prettier-ignore
     if (isSwitchComponent(component)) {
-        component.On();
-        await waiter('On', () => { return component.getValue('state') === true; }, true);
+      component.On();
+      await waiter('On', () => { return component.getValue('state') === true; }, true);
 
-        component.Off();
-        await waiter('Off', () => { return component.getValue('state') === false; }, true);
+      component.Off();
+      await waiter('Off', () => { return component.getValue('state') === false; }, true);
 
-        component.Toggle();
-        await waiter('Toggle', () => { return component.getValue('state') === true; }, true);
+      component.Toggle();
+      await waiter('Toggle', () => { return component.getValue('state') === true; }, true);
 
-        component.Off();
-        await waiter('Off', () => { return component.getValue('state') === false; }, true);
-      }
+      component.Off();
+      await waiter('Off', () => { return component.getValue('state') === false; }, true);
+    }
 
     shelly.removeDevice(device);
     device.destroy();
@@ -423,11 +521,21 @@ describe('Shellies', () => {
     expect(device.getComponentNames()).toStrictEqual(['Ble', 'Cloud', 'Input', 'MQTT', 'Switch', 'Sys', 'Sntp', 'WiFi', 'WS']);
     expect(device.getComponentIds()).toStrictEqual(['ble', 'cloud', 'input:0', 'input:1', 'mqtt', 'switch:0', 'switch:1', 'sys', 'sntp', 'wifi_ap', 'wifi_sta', 'wifi_sta1', 'ws']);
 
-    expect(device.bthomeDevices.size).toBe(1);
+    expect(device.bthomeTrvs.size).toBe(0);
+
+    expect(device.bthomeDevices.size).toBe(4);
     expect(device.bthomeDevices.has('7c:c6:b6:58:b9:a0')).toBe(true);
     expect(device.bthomeDevices.get('7c:c6:b6:58:b9:a0')?.model).toBe('Shelly BLU RC Button 4');
+    expect(device.bthomeDevices.has('0c:ef:f6:01:8d:b8')).toBe(true);
+    expect(device.bthomeDevices.get('0c:ef:f6:01:8d:b8')?.model).toBe('Shelly BLU Wall Switch 4');
+    expect(device.bthomeDevices.has('0c:ef:f6:f1:d7:7b')).toBe(true);
+    expect(device.bthomeDevices.get('0c:ef:f6:f1:d7:7b')?.model).toBe('Shelly BLU DoorWindow');
+    expect(device.bthomeDevices.has('0c:ae:5f:5a:0b:fa')).toBe(true);
+    expect(device.bthomeDevices.get('0c:ae:5f:5a:0b:fa')?.model).toBe('Shelly BLU Motion');
 
-    expect(device.bthomeSensors.size).toBe(5);
+    expect(device.bthomeSensors.size).toBe(18);
+
+    // BLU RC Button 4
     expect(device.bthomeSensors.has('bthomesensor:200')).toBe(true);
     expect(device.bthomeSensors.get('bthomesensor:200')?.addr).toBe('7c:c6:b6:58:b9:a0');
     expect(device.bthomeSensors.get('bthomesensor:200')?.name).toBe('Battery');
@@ -457,7 +565,87 @@ describe('Shellies', () => {
     expect(device.bthomeSensors.get('bthomesensor:204')?.name).toBe('Button');
     expect(device.bthomeSensors.get('bthomesensor:204')?.sensorId).toBe(58);
     expect(device.bthomeSensors.get('bthomesensor:204')?.sensorIdx).toBe(3);
-    expect(device.bthomeTrvs.size).toBe(0);
+
+    // BLU RC Wall Switch 4
+    expect(device.bthomeSensors.has('bthomesensor:205')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:205')?.addr).toBe('0c:ef:f6:01:8d:b8');
+    expect(device.bthomeSensors.get('bthomesensor:205')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:205')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:205')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:206')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:206')?.addr).toBe('0c:ef:f6:01:8d:b8');
+    expect(device.bthomeSensors.get('bthomesensor:206')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:206')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:206')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:207')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:207')?.addr).toBe('0c:ef:f6:01:8d:b8');
+    expect(device.bthomeSensors.get('bthomesensor:207')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:207')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:207')?.sensorIdx).toBe(1);
+
+    expect(device.bthomeSensors.has('bthomesensor:208')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:208')?.addr).toBe('0c:ef:f6:01:8d:b8');
+    expect(device.bthomeSensors.get('bthomesensor:208')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:208')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:208')?.sensorIdx).toBe(2);
+
+    expect(device.bthomeSensors.has('bthomesensor:209')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:209')?.addr).toBe('0c:ef:f6:01:8d:b8');
+    expect(device.bthomeSensors.get('bthomesensor:209')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:209')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:209')?.sensorIdx).toBe(3);
+
+    // BLU DoorWindow
+    expect(device.bthomeSensors.has('bthomesensor:210')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:210')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:210')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:210')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:210')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:211')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:211')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:211')?.name).toBe('Illuminance');
+    expect(device.bthomeSensors.get('bthomesensor:211')?.sensorId).toBe(5);
+    expect(device.bthomeSensors.get('bthomesensor:211')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:212')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:212')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:212')?.name).toBe('Contact');
+    expect(device.bthomeSensors.get('bthomesensor:212')?.sensorId).toBe(45);
+    expect(device.bthomeSensors.get('bthomesensor:212')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:213')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:213')?.addr).toBe('0c:ef:f6:f1:d7:7b');
+    expect(device.bthomeSensors.get('bthomesensor:213')?.name).toBe('Rotation');
+    expect(device.bthomeSensors.get('bthomesensor:213')?.sensorId).toBe(63);
+    expect(device.bthomeSensors.get('bthomesensor:213')?.sensorIdx).toBe(0);
+
+    // BLU Motion
+    expect(device.bthomeSensors.has('bthomesensor:214')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:214')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:214')?.name).toBe('Battery');
+    expect(device.bthomeSensors.get('bthomesensor:214')?.sensorId).toBe(1);
+    expect(device.bthomeSensors.get('bthomesensor:214')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:215')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:215')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:215')?.name).toBe('Motion');
+    expect(device.bthomeSensors.get('bthomesensor:215')?.sensorId).toBe(33);
+    expect(device.bthomeSensors.get('bthomesensor:215')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:216')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:216')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:216')?.name).toBe('Button');
+    expect(device.bthomeSensors.get('bthomesensor:216')?.sensorId).toBe(58);
+    expect(device.bthomeSensors.get('bthomesensor:216')?.sensorIdx).toBe(0);
+
+    expect(device.bthomeSensors.has('bthomesensor:217')).toBe(true);
+    expect(device.bthomeSensors.get('bthomesensor:217')?.addr).toBe('0c:ae:5f:5a:0b:fa');
+    expect(device.bthomeSensors.get('bthomesensor:217')?.name).toBe('Illuminance');
+    expect(device.bthomeSensors.get('bthomesensor:217')?.sensorId).toBe(5);
+    expect(device.bthomeSensors.get('bthomesensor:217')?.sensorIdx).toBe(0);
 
     const component = device.getComponent('switch:0');
     expect(component).not.toBeUndefined();